+++ src/app/page.tsx
@@ -1,8 +1,8 @@
-
+import Navbar from '@/components/presentation/Navbar';
 export default function Home() {
   return (
     <div>
-  
+      <Navbar />
     </div>
   );
 }
